Changes names and removes redundant functions
